services/parser_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
python-docx
easyocr
